Major overhaul
Removed OnStartConstraint and OnFinishConstraint. Replaced with CreateConstraint which reduces much of the copypasted code through the script.
Added a check for valid constraint in OnConstraintRemoved as it may be called more than once on the same entity.
Small optimizations to RemoveAll and Find.
Optimizations and cleanup to all constraint-generating functions.
Made OnRemoveConstraint and CreateConstraint global (Necessary for custom constraints to be made).
New GetTableStyle function to return a GetTable-styled table from a single constraint
Small optimization to GetTable
Heavily optimized FindConstraint, FindConstraint, and FindConstraintEntity